+++ src/js/bootstrapValidator.js
@@ -735,8 +735,7 @@
                             // v is validator name
                             $f.removeData('bv.dfs.' + v);
                             if (message) {
-                                // Update the error message
-                                $field.data('bv.messages').find('.help-block[data-bv-validator="' + v + '"][data-bv-for="' + $f.attr('data-bv-field') + '"]').html(message);
+                                that.updateMessage($f, v, message);
                             }
 
                             that.updateStatus(updateAll ? $f.attr('data-bv-field') : $f, isValid ? that.STATUS_VALID : that.STATUS_INVALID, v);
@@ -1057,9 +1056,11 @@
          *
          * @param {String|jQuery} [field] The field name or field element
          * If the field is not defined, the method returns all error messages of all fields
+         * @param {String} [validator] The name of validator
+         * If the validator is not defined, the method returns error messages of all validators
          * @returns {String[]}
          */
-        getMessages: function(field) {
+        getMessages: function(field, validator) {
             var that     = this,
                 messages = [],
                 $fields  = $([]);
@@ -1080,11 +1081,12 @@
                     break;
             }
 
+            var filter = validator ? '[data-bv-validator="' + validator + '"]' : '';
             $fields.each(function() {
                 messages = messages.concat(
                     $(this)
                         .data('bv.messages')
-                        .find('.help-block[data-bv-for="' + $(this).attr('data-bv-field') + '"][data-bv-result="' + that.STATUS_INVALID + '"]')
+                        .find('.help-block[data-bv-for="' + $(this).attr('data-bv-field') + '"][data-bv-result="' + that.STATUS_INVALID + '"]' + filter)
                         .map(function() {
                             return $(this).html();
                         })
@@ -1093,6 +1095,22 @@
             });
 
             return messages;
+        },
+
+        /**
+         * Update the error message
+         *
+         * @param {String|jQuery} field The field name or field element
+         * @param {String} validator The validator name
+         * @param {String} message The message
+         * @returns {BootstrapValidator}
+         */
+        updateMessage: function(field, validator, message) {
+            var fields = ('string' === typeof field) ? this.getFieldElements(field) : field,
+                field  = ('object' === typeof field) ? field.attr('data-bv-field')  : field;
+            fields.each(function() {
+                $(this).data('bv.messages').find('.help-block[data-bv-validator="' + validator + '"][data-bv-for="' + field + '"]').html(message);
+            });
         },
 
         /**